+++ CHANGES.md
@@ -2,9 +2,12 @@
 
 - `location.query` has no prototype
 - Warn about protocol-relative URLs ([#243])
+- **Bugfix:** Ignore errors when saving hash history state if
+  `window.sessionStorage` is undefined ([#295])
 
 [HEAD]: https://github.com/mjackson/history/compare/v3.0.0-2...HEAD
 [#243]: https://github.com/mjackson/history/issues/243
+[#295]: https://github.com/mjackson/history/issues/295
 
 ## [v3.0.0-2]
 > Apr 19, 2016